requirements.txt
psycopg2
psycopg2-binary
requests
uuid
faker
fastapi
pydantic
starlette
supabase
jinja2
uvicorn
python-multipart
python-jose[cryptography]

python-dotenv
SQLAlchemy
pydantic[email]